# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RailYatra API
  version: 2.0.0
  description: |
    Node.js + SQL Server API for RailYatra train booking and reservation management.
    Master timetable data (Category A) is imported from authorized datasets.
    Fares, availability, RAC/WL (Category B) are development simulations unless noted.
servers:
  - url: /api
paths:
...
  /trains/{id}/route:
    get:
      summary: Ordered train route with day offsets
      tags: [Trains]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Stops sorted by stopSequence
...
  /bookings/pnr/{pnr}:
    get:
      summary: PNR status lookup
      tags: [Bookings]
      parameters:
        - { name: pnr, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Booking and passenger status (PII limited)
        '404':
          description: PNR not found
  /bookings/{id}/pending:
    delete:
      summary: Remove unpaid pending booking
      tags: [Bookings]
      security: [{ AuthToken: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Pending booking removed
  /passengers/saved:
    get:
      summary: List saved passengers for current user
      tags: [Passengers]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: Saved passenger list
    post:
      summary: Add saved passenger
      tags: [Passengers]
      security: [{ AuthToken: [] }]
      responses:
        '201':
          description: Saved passenger created
  /passengers/saved/{id}:
    put:
      summary: Update saved passenger
      tags: [Passengers]
      security: [{ AuthToken: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Updated passenger
    delete:
      summary: Delete saved passenger
      tags: [Passengers]
      security: [{ AuthToken: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Passenger removed
  /payments/create-order:
    post:
      summary: Create Razorpay order for booking payment
      tags: [Payments]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: Payment order details
  /payments/verify:
    post:
      summary: Verify Razorpay payment signature
      tags: [Payments]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: Payment confirmed
  /payments/dev-confirm:
    post:
      summary: Simulate payment (dev mode only)
      tags: [Payments]
      security: [{ AuthToken: [] }]
      responses:
        '200':
          description: Payment simulated
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: x-auth-token